add: crud master user by admin
